D & D Character Roller
This was designed for personal use only for quick generation of
character sheets for the game Dungeons And Dragons.  It was submitted as
a final test program for CIT 195 at NMC (Intro to C#).  Though the
program is unfinished, it shows proficient use of OOD in C# as well as
usage of common controls.